+++ geotrek/feedback/tests/test_views.py
@@ -16,7 +16,7 @@ from geotrek.maintenance.tests.factories import InfrastructureInterventionFactor
 from mapentity.tests.factories import SuperUserFactory, UserFactory
 from rest_framework.test import APIClient
 
-from geotrek.common.tests import CommonTest, TranslationResetMixin
+from geotrek.common.tests import CommonTest, TranslationResetMixin, GeotrekAPITestCase
 from geotrek.common.utils.testdata import (get_dummy_uploaded_file,
                                            get_dummy_uploaded_image,
                                            get_dummy_uploaded_image_svg)
@@ -100,7 +100,7 @@ class ReportSerializationOptimizeTests(TestCase):
         self.filed_report = feedback_factories.ReportFactory(status=self.filed_status)
 
 
-class ReportViewsTest(CommonTest):
+class ReportViewsTest(GeotrekAPITestCase, CommonTest):
     model = feedback_models.Report
     modelfactory = feedback_factories.ReportFactory
     userfactory = SuperUserFactory
@@ -119,6 +119,16 @@ class ReportViewsTest(CommonTest):
             'email': self.obj.email,
             'status': self.obj.status_id,
             'problem_magnitude': self.obj.problem_magnitude.pk
+        }
+
+    def get_expected_datatables_attrs(self):
+        return {
+            'activity': self.obj.activity.label,
+            'category': self.obj.category.label,
+            'date_update': '17/03/2020 00:00:00',
+            'email': self.obj.email_display,
+            'id': self.obj.pk,
+            'status': self.obj.status
         }
 
     def get_bad_data(self):
@@ -225,7 +235,7 @@ class CreateReportsAPITest(BaseAPITest):
         client = APIClient()
         response = client.post(self.add_url, data=data,
                                allow_redirects=False)
-        self.assertEqual(response.status_code, 201)
+        self.assertEqual(response.status_code, 201, self.add_url)
 
     def test_reports_can_be_created_using_post(self):
         self.post_report_data(self.data)