+++ geotrek/feedback/views.py
@@ -1,4 +1,3 @@
-
 from django.conf import settings
 from django.contrib.auth import get_user_model
 from django.contrib.contenttypes.models import ContentType
@@ -6,18 +5,21 @@ from django.core.mail import send_mail
 from django.urls.base import reverse
 from django.utils.translation import gettext as _
 from django.views.generic.list import ListView
-from geotrek.common.mixins import CustomColumnsMixin
-from geotrek.common.models import Attachment, FileType
-from geotrek.feedback import models as feedback_models
-from geotrek.feedback import serializers as feedback_serializers
-from geotrek.feedback.filters import ReportFilterSet
-from geotrek.feedback.forms import ReportForm
 from mapentity import views as mapentity_views
 from rest_framework.decorators import action
 from rest_framework.parsers import FormParser, MultiPartParser
 from rest_framework.permissions import AllowAny
 from rest_framework.response import Response
 from rest_framework.views import APIView
+
+from geotrek.common.mixins.api import APIViewSet
+from geotrek.common.mixins.views import CustomColumnsMixin
+from geotrek.common.models import Attachment, FileType
+from geotrek.common.viewsets import GeotrekMapentityViewSet
+from geotrek.feedback import models as feedback_models
+from geotrek.feedback import serializers as feedback_serializers
+from geotrek.feedback.filters import ReportFilterSet
+from geotrek.feedback.forms import ReportForm
 
 
 class ReportLayer(mapentity_views.MapEntityLayer):
@@ -70,6 +72,7 @@ class ReportList(CustomColumnsMixin, mapentity_views.MapEntityList):
     filterform = ReportFilterSet
     mandatory_columns = ['id', 'email', 'activity']
     default_extra_columns = ['category', 'status', 'date_update']
+    searchable_columns = ['id', 'email']
 
     def get_queryset(self):
         qs = super().get_queryset()  # Filtered by FilterSet
@@ -141,20 +144,31 @@ class ReportUpdate(mapentity_views.MapEntityUpdate):
     form_class = ReportForm
 
 
-class ReportViewSet(mapentity_views.MapEntityViewSet):
+class ReportViewSet(GeotrekMapentityViewSet):
     """Disable permissions requirement"""
 
     model = feedback_models.Report
-    queryset = (
-        feedback_models.Report.objects.existing()
-        .select_related(
-            "activity", "category", "problem_magnitude", "status", "related_trek"
-        )
-        .prefetch_related("attachments")
-    )
-    parser_classes = [FormParser, MultiPartParser]
     serializer_class = feedback_serializers.ReportSerializer
-    geojson_serializer_class = feedback_serializers.ReportGeojsonSerializer
+    authentication_classes = []
+    permission_classes = [AllowAny]
+
+    def get_columns(self):
+        return ReportList.mandatory_columns + settings.COLUMNS_LISTS.get('feedback_view',
+                                                                         ReportList.default_extra_columns)
+
+    def get_queryset(self):
+        return self.model.objects.existing().select_related(
+            "activity", "category", "problem_magnitude", "status", "related_trek"
+        ).prefetch_related("attachments")
+
+
+class ReportAPIViewSet(APIViewSet):
+    queryset = feedback_models.Report.objects.existing()\
+                              .select_related("activity", "category", "problem_magnitude", "status", "related_trek")\
+                              .prefetch_related("attachments")
+    parser_classes = [FormParser, MultiPartParser]
+    serializer_class = feedback_serializers.ReportAPISerializer
+    geojson_serializer_class = feedback_serializers.ReportAPIGeojsonSerializer
     authentication_classes = []
     permission_classes = [AllowAny]