+++ geotrek/maintenance/views.py
@@ -1,25 +1,23 @@
-
 import logging
 
 from django.conf import settings
 from django.db.models import Subquery, OuterRef, Sum
 from django.db.models.expressions import Value
 from django.utils.translation import gettext_lazy as _
-from mapentity.views import (MapEntityLayer, MapEntityList, MapEntityJsonList, MapEntityFormat, MapEntityViewSet,
-                             MapEntityDetail, MapEntityDocument, MapEntityCreate, MapEntityUpdate, MapEntityDelete)
+from mapentity.views import (MapEntityLayer, MapEntityList, MapEntityFormat, MapEntityDetail, MapEntityDocument,
+                             MapEntityCreate, MapEntityUpdate, MapEntityDelete)
 
 from geotrek.altimetry.models import AltimetryMixin
-from geotrek.common.mixins import CustomColumnsMixin
-from geotrek.common.views import FormsetMixin
 from geotrek.authent.decorators import same_structure_required
-from .models import Intervention, Project, ManDay
+from geotrek.common.mixins.forms import FormsetMixin
+from geotrek.common.mixins.views import CustomColumnsMixin
+from geotrek.common.viewsets import GeotrekMapentityViewSet
 from .filters import InterventionFilterSet, ProjectFilterSet
 from .forms import (InterventionForm, ProjectForm,
                     FundingFormSet, ManDayFormSet)
+from .models import Intervention, Project, ManDay
 from .serializers import (InterventionSerializer, ProjectSerializer,
                           InterventionGeojsonSerializer, ProjectGeojsonSerializer)
-from rest_framework import permissions as rest_permissions
-
 
 logger = logging.getLogger(__name__)
 
@@ -35,10 +33,7 @@ class InterventionList(CustomColumnsMixin, MapEntityList):
     filterform = InterventionFilterSet
     mandatory_columns = ['id', 'name']
     default_extra_columns = ['date', 'type', 'target', 'status', 'stake']
-
-
-class InterventionJsonList(MapEntityJsonList, InterventionList):
-    pass
+    searchable_columns = ['id', 'name']
 
 
 class InterventionFormatList(MapEntityFormat, InterventionList):
@@ -168,16 +163,17 @@ class InterventionDelete(MapEntityDelete):
         return super().dispatch(*args, **kwargs)
 
 
-class InterventionViewSet(MapEntityViewSet):
+class InterventionViewSet(GeotrekMapentityViewSet):
     model = Intervention
-    queryset = Intervention.objects.existing()
     serializer_class = InterventionSerializer
     geojson_serializer_class = InterventionGeojsonSerializer
-    permission_classes = [rest_permissions.DjangoModelPermissionsOrAnonReadOnly]
+    filterset_class = InterventionFilterSet
 
     def get_queryset(self):
-        # Override annotation done by MapEntityViewSet.get_queryset()
-        return Intervention.objects.all()
+        return self.model.objects.existing()
+
+    def get_columns(self):
+        return InterventionList.mandatory_columns + settings.COLUMNS_LISTS.get('intervention_view', InterventionList.default_extra_columns)
 
 
 class ProjectLayer(MapEntityLayer):
@@ -194,10 +190,7 @@ class ProjectList(CustomColumnsMixin, MapEntityList):
     filterform = ProjectFilterSet
     mandatory_columns = ['id', 'name']
     default_extra_columns = ['period', 'type', 'domain']
-
-
-class ProjectJsonList(MapEntityJsonList, ProjectList):
-    pass
+    searchable_columns = ['id', 'name']
 
 
 class ProjectFormatList(MapEntityFormat, ProjectList):
@@ -252,13 +245,14 @@ class ProjectDelete(MapEntityDelete):
         return super().dispatch(*args, **kwargs)
 
 
-class ProjectViewSet(MapEntityViewSet):
+class ProjectViewSet(GeotrekMapentityViewSet):
     model = Project
-    queryset = Project.objects.existing()
     serializer_class = ProjectSerializer
     geojson_serializer_class = ProjectGeojsonSerializer
-    permission_classes = [rest_permissions.DjangoModelPermissionsOrAnonReadOnly]
+    filterset_class = ProjectFilterSet
 
     def get_queryset(self):
-        # Override annotation done by MapEntityViewSet.get_queryset()
-        return Project.objects.all()
+        return self.model.objects.existing()
+
+    def get_columns(self):
+        return ProjectList.mandatory_columns + settings.COLUMNS_LISTS.get('project_view', ProjectList.default_extra_columns)