+++ geotrek/trekking/urls.py
@@ -2,6 +2,7 @@ from django.conf import settings
 from django.urls import path, register_converter
 
 from mapentity.registry import registry
+from rest_framework.routers import DefaultRouter
 
 from geotrek.altimetry.urls import AltimetryEntityOptions
 from geotrek.common.urls import PublishableEntityOptions, LangConverter
@@ -11,7 +12,7 @@ from . import models
 from .views import (
     TrekDocumentPublic, TrekDocumentBookletPublic, TrekMapImage, TrekMarkupPublic,
     TrekGPXDetail, TrekKMLDetail, WebLinkCreatePopup, TrekPOIViewSet,
-    TrekServiceViewSet
+    TrekServiceViewSet, TrekAPIViewSet, POIAPIViewSet, ServiceAPIViewSet
 )
 
 register_converter(LangConverter, 'lang')
@@ -49,6 +50,14 @@ class ServiceEntityOptions(MapEntityOptions):
     pass
 
 
+router = DefaultRouter(trailing_slash=False)
+
+
+router.register(r'^api/(?P<lang>[a-z]{2})/treks', TrekAPIViewSet, basename='trek')
+router.register(r'^api/(?P<lang>[a-z]{2})/pois', POIAPIViewSet, basename='poi')
+router.register(r'^api/(?P<lang>[a-z]{2})/services', ServiceAPIViewSet, basename='service')
+
+urlpatterns += router.urls
 urlpatterns += registry.register(models.Trek, TrekEntityOptions, menu=settings.TREKKING_MODEL_ENABLED)
 urlpatterns += registry.register(models.POI, POIEntityOptions, menu=settings.POI_MODEL_ENABLED)
 urlpatterns += registry.register(models.Service, ServiceEntityOptions, menu=settings.SERVICE_MODEL_ENABLED)